fix my thinko; elminate remaining [gcc 12.2.1] compiler warnings
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/copula@1882 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/copula/src/rF01Joe.c
@@ -56,12 +56,12 @@ double rF01Joe(double V0, double alpha, double gamma_1_a /**< == Gamma(1 - alpha
  * @return none
  * @author Marius Hofert
  */
-void rF01Joe_vec(double* V01, const double *V0, R_xlen_t n, double alpha, double approx){
+void rF01Joe_vec(double* V01, const double *V0, R_xlen_t n, double alpha, double approx) {
     double gamma_1_a = gammafn(1. - alpha);
     GetRNGstate();
 
     for(R_xlen_t i=0; i < n; i++)
-	V01[i] = rF01Joe(V0[i], alpha, gamma_1_a, approx);
+	V01[i] = rF01Joe(V0[i], alpha, gamma_1_a, (int)approx);
 
     PutRNGstate();
     return;
